+++ gaios/UserSettings/Cells/WordCell.swift
@@ -9,8 +9,5 @@ class WordCell: UICollectionViewCell {
     func configure(num: Int, word: String) {
         self.lblWord.text = word
         self.lblNum.text = "\(num + 1)"
-        bg.borderWidth = 1.0
-        bg.borderColor = .white.withAlphaComponent(0.2)
-        bg.layer.cornerRadius = 3.0
     }
 }